[FIX] only aggregate columns explicitly asking for aggregation                                                                                                ~/projects/tiny/web/current
Instead of relying only on the column's group_operator, ensure the
view itself asks for the aggregation function (by checking if the
corresponding field attribute is set).

If the @group_operator function is not set on the field (even without
a label, it just has to be *there* so `<field sum=""/>` would work),
do not aggregate the column at all.