[FIX] ensure o2m widget has saved its sub-views before it returns its value to its parent form view
This is currently highly faillible as saving a form or a list view can
be an async operation, yet the FormView's `Widget#get_value` call can
only be synchronous. As a result, there are cases where saving a form
view can fail to get the correct value for its o2m fields.

Ideally, `Widget#get_value` should be an async API (or a sync/async API)