[FIX] avoid systematically deadlocking the stand-alone OpenERP web in its default configuration
The default configuration for OpenERP Web standalone (launched via the
openerp-web script) is single-threaded for log-readability purposes.

web.common.controllers.main.Proxy.load has been added to wrap a
request to an HTTP handler inside a JSON (and/or JSONP) request, the
initial implementation was to perform a full HTTP call *from within an
HTTP handler to itself*.

Since the server is single-threaded and its only thread is already
busy, it can't handle the new request, and the client deadlocks.

Replaced this crap by also-crap-but-slightly-less-so: instantiating a
Werkzeug test client using the root application and proxying the
request through that. Avoids creating a new request from the server,
therefore does not deadlock.