[FIX] double-include of ids in the dataset, and double-save in some cases                                                                                                       ~/projects/tiny/web/current
on [change] of a listview record, if there was no id on the record
(id=false) we need to add one. But depending how the creation/addition
of the record was done, the id may already have been inserted e.g. by
the formview.

So first check if the record's id is already in the dataset, and add
it if it's not.