[FIX] fetching of table cells to match form inputs to
Only fetch the table cells of the row being edited *after* having
saved a pending edition.

Otherwise, when re-editing the same record (e.g. click between buttons
or on a readonly field), the cells are saved, the record is refreshed
destroying the cells, and when trying to get the cell positions to
position form fields correctly... the cells which don't exist anymore
return 0, 0 as their position and nothing works correctly.