[FIX] tabbing out of an m2o which opens some sort of weird dialog thing while in an editable o2m
The tab blurs the current fields, focuses the next one, then the
dialog thing opens and steals the focus from the *next* field (not the
one which created it), which thus triggers *blur* ultimately blurring
(and saving) the row being edited.

Add a trigger call in some random place which seems to make things not
fail.