[FIX] correctly handle case of deleting-an-o2m-line-being-edited
move special case in o2m listview to main editable and expand to
correctly fit edition cases, not just work in case of creation

issue: if a row being edited is deleted in an o2m, the form will be
blurred and try cancelling/saving the edition and reloading the row,
but this will only happen *after* the record/row/whatever has been
"deleted" and (amongst other things) removed from the buffered o2m
cache, leading to a tentative read on the server side (usually with a
local id which doesn't even make sense) and thus an error.

Intercept the deletion signal to cancel the edition immediately and
only let the deletion through after the edition has successfully been
canceled.