[FIX] pasting HTML content into the text box, then trying to navigate/alter it
Navigation implementation can only deal with straight text (and
asserts that), if HTML is pasted in a search input
InputView#getSelection will throw errors and refuse to act.

Clean up input content after a paste event, to ensure only plain text
is present so it can be navigated.

Don't forget to correctly re-set the cursor at the end of the input
data, otherwise the user will face various deep DOM errors when trying
to move around the input with the arrow keys (which he would usually
be able to do after a paste).