[FIX] cleanup m2o field events when reloading m2o fields
The big issue was the hooking on the closest ui-dialog's scroll event:
if an m2o in a dialog (e.g. wizard-type situations) toggled state from
editable to readonly (workflow action of some sort), the event hook
would remain, and call .$input.autocomplete("widget") which would blow
up because the autocompletion widget hasn't been re-initialized.

Take care to fully unbind all events during the reload process to
avoid this issue (and potentially other such as duplicate bindings and
the like).