[FIX] replace uncommon month names of the ar_SY datejs localization file by more standard ones, courtesy of Ahmad Khayyat
The server only supports the ar_SY variant of arabic, so this is the
one and only locale arabic speakers can currently select. According to
Ahmad Khayyat:

> The [ar-SY datejs localization] uses uncommon month names, which are
> hard to recognize.  The more common month names should be used
> instead.

This could be reverted if more arabic locales are made available in
the server, in order to limit deviations from the upstream package.