[FIX] broken pager
Turns out when code looks somewhat odd there may well be a good reason
for it, and changing it without wondering breaks the pager.

In this case, `/web/dataset/search_read` has a significant difference
with Model.search_read: it returns the records slice specified by
(``limit``, ``offset``) but it also returns the *total number of
records* for ``domain`` which is sort-of useful to generating the
pager.