[FIX] close autocomplete listing on input
Closing previously occured on search request (so that a user wouldn't be able
to select "old" data on new search request), but ``search`` is only triggered
after the search delay. Worked when delay was 0, with it being moved to 250 a
user can get results matching the previous search instead of the current one.

Trigger a closing of the current results list on any ``input`` event, which is
when text is entered in any of the searchview's ``InputView``