[FIX] Redirect user from /web if a get redirect is found.
If you redirect an user not logged, into /web?redirect=xxx, he was correctly redirect to /web/login?redirect=/web?redirect=xxx. But when user come back on /web?redirect=xxx, he was not redirected to xxx.

The question should be to have only one behavior when we redirect people, as well /web?redirect as well /web/login?redirect. But because we use both in actual code, we need to accept both for retro-compatibility.