[FIX] (partially) fixes a concurrence problem in graph view
The graph view is asynchronous when updating data.  It is not safe when 
the user tries to update quickly the group bys (for example, when adding two 
group bys in quick succession).  

This patch partially fixes the problem: it makes sure that two concurrent updates 
will be correctly serialized.  However, it is not a complete fix: the crash can still happen
when three or more updates are quickly done.  A complete solution require some more work in 
keeping tracks of an update queue and serializing properly the updates. (will be done in trunk)