[IMP] large optimizations to the graph view rendering (pivot table)
The pivot table rendering was slow with large tables : it appended to the
table each line, so the browser had to rerender each time the page.  This
was solved by rendering the table to a document fragment, then only add
the fragment to the main document.  This speeds up considerably the process
(from 2600ms to 420ms in one test)

Also, uses basic loops instead of underscore maps in the build_rows method